% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster.l2fc}
\alias{cluster.l2fc}
\title{cluster.l2fc
cluster data based on l2fc}
\usage{
cluster.l2fc(mat, data_l2fc)
}
\arguments{
\item{mat}{dataset}

\item{data_l2fc}{known l2fc}
}
\value{
list of matrices
}
\description{
cluster.l2fc
cluster data based on l2fc
}
