% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{compute.l2fc}
\alias{compute.l2fc}
\title{compute.l2fc
calculate l2fc between clusters}
\usage{
compute.l2fc(data, labels)
}
\arguments{
\item{data}{dataset (transposed)}

\item{labels}{cluster label values}
}
\value{
vector
}
\description{
compute.l2fc
calculate l2fc between clusters
}
