% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_data}
\alias{count_data}
\title{Example gene expression count data.}
\format{
An object of class \code{data.frame} with 882 rows and 348 columns.
}
\usage{
count_data
}
\description{
Example dataset of gene expression data (882 genes) from
various cell lines (348 samples).
}
\keyword{datasets}
