% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist.corr}
\alias{dist.corr}
\title{dist.corr
calculate the distance between two datasets or distance of one dataset, if d2 is null}
\usage{
dist.corr(
  data,
  d2 = NULL,
  method = c("pearson", "kendall", "spearman", "euclidean", "manhattan")
)
}
\arguments{
\item{data}{first dataset}

\item{d2}{second dataset}

\item{method}{method of distance calculation}
}
\value{
distance matrix
}
\description{
dist.corr
calculate the distance between two datasets or distance of one dataset, if d2 is null
}
