% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaclust.R
\name{ga.clust}
\alias{ga.clust}
\title{ga.clust
Main function, calls GA function to cluster data using correlation-based fitness}
\usage{
ga.clust(
  dataset = NULL,
  crossover.rate = 0.9,
  k = 2,
  mutation.rate = 0.01,
  elitism = 0.05,
  pop.size = 25,
  generations = 100,
  seed.ga = 42,
  method = "pearson",
  penalty.function = NULL,
  plot.internals = TRUE,
  known_lfc = NULL,
  ...
)
}
\arguments{
\item{dataset}{dataset to cluster}

\item{crossover.rate}{probability crossover occurs}

\item{k}{number of clusters}

\item{mutation.rate}{probability of mutation}

\item{elitism}{\% of population retained in next iter}

\item{pop.size}{initial population size}

\item{generations}{repetitions}

\item{seed.ga}{seed}

\item{method}{correlation method}

\item{penalty.function}{fitness penalty}

\item{plot.internals}{plot internal messages?}

\item{known_lfc}{external lfc exists?}

\item{...}{additional params}
}
\value{
ga.clust object
}
\description{
ga.clust
Main function, calls GA function to cluster data using correlation-based fitness
}
