% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input_lfc}
\alias{input_lfc}
\title{DESeq2 Calculated Log2FoldChange}
\format{
An object of class \code{numeric} of length 882.
}
\usage{
input_lfc
}
\description{
A numeric vector containing Log2FoldChange calculations of
gene expression count data. This data takes the L2FC of
count_data.
}
\keyword{datasets}
