% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t_count_data}
\alias{t_count_data}
\title{Transposed version of count_data.}
\format{
An object of class \code{data.frame} with 348 rows and 882 columns.
}
\usage{
t_count_data
}
\description{
Example dataset of gene expression data (882 genes) from
various cell lines (348 samples).
}
\keyword{datasets}
